caveman/requirements.txt
spacy
networkx
tiktoken
openai
transformers
torch
python-dotenv